iron_man/requirements.txt
certifi==2020.6.20 
chardet==3.0.4 
click==7.1.2 
Flask==1.1.2
future==0.18.2 
gunicorn==20.0.4 
idna==2.10 
itsdangerous==1.1.0
Jinja2==2.11.2 
line-bot-sdk==1.8.0 
MarkupSafe==1.1.1 
requests==2.24.0
urllib3==1.25.10 
Werkzeug==1.0.1 
wincertstore==0.2
flask_apscheduler
pymongo==3.12.3
pymongo[srv]==3.12.3
schedule
APScheduler